Relocated PromptForm component
- Removed unnecessary vite.config.js comment
