+++ frontend/src/components/PromptForm/PromptForm.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
-import Button from "./Button/Button";
-import Card from "./Card/Card";
-import { generateSuggestions } from "../services/api";
+import Button from "../Button/Button";
+import Card from "../Card/Card";
+import { generateSuggestions } from "../../services/api";
 
 interface InputCardFormProps {
   isBackendWorking: boolean;
