+++ frontend/vite.config.js
@@ -1,6 +1,5 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
-// import { defineConfig } from "vitest";
 
 // https://vitejs.dev/config/
 export default defineConfig({
